Finally, all 256 interrupts are usable
Till now, only 255 were possible because the vector was an 8-bit value instead of 16, which basically meant an overflow and the IDT would reload gates all over again

+++ src/include/idt.c
@@ -30,7 +30,8 @@ void idt_init() {
     idtr.base = (idt_entry_t*)&idt_entries[0];
     idtr.limit = (uint16_t)sizeof(idt_entry_t) * IDT_MAX_DESCRIPTORS - 1;
 
-    for (uint8_t vector = 0; vector < IDT_MAX_DESCRIPTORS; vector++) {
+    for (uint16_t vector = 0; vector < IDT_MAX_DESCRIPTORS; vector++) {
+        // printf("Enabling gate %d\n", vector);
         idt_set_gate(vector, isr_stub_table[vector], GDT_CODE_SEGMENT, 0x8E);
         vectors[vector] = true;
     }
